glloader: Add new extensions. KlayGE: OpenGL & OpenGLES: Remove warnings.

+++ KlayGE/Plugins/Src/Render/OpenGL/OGLRenderEngine.cpp
@@ -1454,7 +1454,7 @@ namespace KlayGE
 		}
 
 		glGetIntegerv(GL_MAX_VERTEX_ATTRIBS, &temp);
-		caps_.max_vertex_streams = temp;
+		caps_.max_vertex_streams = static_cast<uint8_t>(temp);
 
 		caps_.hw_instancing_support = true;
 		caps_.instance_id_support = false;
